requirements.txt
hdx-python-api~=5.9.0
geopandas~=0.10.2
Fiona~=1.8.20
pandas~=1.3.4
Shapely~=1.8.0
